nicer repl
